Refactor common code 'KurtosisNode
